Fixing to use with postman

+++ API/deleteContact.php
@@ -2,22 +2,17 @@
 include "../conn.php";
 session_start();
 
-if (isset($_SESSION['id'])) {
-  $user_id = $_SESSION['id'];
-}
-else {
-  header('Location: ../index.php');
-}
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+  $contact_id = $_GET["contact_id"];
 
-$contact_id = $_GET["contact_id"];
-
-$sql = "DELETE FROM `contacts` WHERE `id` = '$contact_id'";
-$result = mysqli_query($conn, $sql);
-
-if ($result) {
-  header("Location: ../dashboard.php");
-} else {
-  echo "Failed: " . mysqli_error($conn);
-  header("Location: ../dashboard.php");
+  $sql = "DELETE FROM `contacts` WHERE `id` = '$contact_id'";
+  $result = mysqli_query($conn, $sql);
+  
+  if ($result) {
+    header("Location: ../dashboard.php");
+  } else {
+    echo "Failed: " . mysqli_error($conn);
+    header("Location: ../dashboard.php");
+  }
 }
 ?>
